requirements.txt
google.generativeai
streamlit
pytesseract
opencv-python
Pillow
python-docx
pdf2image
